requirements.txt
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
cloudpickle==3.0.0
filelock==3.13.1
fsspec==2023.10.0
ftfy==6.1.3
gym==0.26.2
gym-notices==0.0.8
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
importlib-metadata==6.8.0
joblib==1.3.2
lazy_loader==0.3
networkx==3.2.1
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.7.0.68
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.13.3
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.66.1
transformers==4.28.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
